scripts/requirements-migration.txt
# Migration Dependencies for Portfolio Data Processing

# Core database and vector store libraries
psycopg2-binary>=2.9.5
upstash-vector>=0.4.0

# Environment and configuration
python-dotenv>=1.0.0

# Data processing and validation  
pydantic>=2.0.0
jsonschema>=4.17.0

# HTTP client for API calls
requests>=2.31.0

# Logging and progress tracking
tqdm>=4.65.0
colorlog>=6.7.0

# Date/time handling
python-dateutil>=2.8.2

# Optional utilities for data analysis
pandas>=2.0.0
numpy>=1.24.0